Rename transmart-sorting.ts file
