+++ src/app/models/transmart-models/transmart-data-table.ts
@@ -1,6 +1,6 @@
 import {TransmartDimension} from './transmart-dimension';
 import {TransmartRow} from 'app/models/transmart-models/transmart-row';
-import {TransmartSort} from './transmart-sorting';
+import {TransmartSort} from './transmart-sort';
 import {TransmartColumnHeaders} from './transmart-column-headers';
 
 export class TransmartDataTable {
